v0.045d - btc_processor + other(s)
This is a minimal structural update for a visual idea that wasn't my own by my brother's at the heart of it. (omnipotententity/0mn1p0t3nt3nt1ty) aka "0mni" the starbound dev. The Lantern Core is of my own design however the entirety of the BTC_Processor was my brothers suggestion, at least in spirit. So I'm more than happy to mention him as a co-author on this particular aspect of my files. (Visuals and actual working API will be my next main focus to allow TESTING of the system and analysis of the files stores ect, as intended by the heart of the core functionality I am aiming for.

ps. as a dev I may be working under this branch (0.045x to consider working on minor bugs ect until 0.046 which I intend for a historical full integration of some of my first logic AFTER I confirm core functionality. Any BASIC files you see here are my intended launchpad for running this logically on Mac OS, Linux OS, and Windows App

Internal testing will happen as time allows on Mac OS, and will likely be the last OS I am to get this working on.

Essentially. "Mostly" Gaps now just need filled inside the file architecture. Rigorous Testing. And maybe about 2-3 weeks of a discerning eye once I have everything up and running myself.

For now, I believe I'm actually fairly close to a "working" version that is meant for testing only. If you're looking for specifics about update details, I decided it best to casually explain this, as this is getting closer to a production grade version. IMO