0.046 - Visuals & BTC_Processor
This is mostly a lot, so I'm just gonna say
"I did like 150k lines of refactoring over a day or two and a lot of other work and it was a lot."

But we got it done man.

Idk, it felt like a lot of test passed, but that might be the empty shell work before the links to the real visuals and other working aspects like actual functionality. I'm just being very cautions. I'll keep working on this internal system moving forward to make sure that everything works.