v0.046e - new math logic
did a little work trying to see if icould get flake8 working, its been the better part of a day. And there was progress IMO, only

I think we are doing "ok" very obviously still a lot of implementation logic in our main math pipes that need better handling and hand off.

Refactoring was required, and this was my attempt at that.