v0.046h - stub fixes
The goal of this is to improve Overall functionality and reduced flake8 errors. As many as I can address in as short of a time as I can do it.

I'm also working on the structure of The overall trade orchestration, just in case. I call it ghost, but it's pretty straightforward.