v0.049W - winning
I think most of the CLI works, and I worked on a little bit of the CUDA acceleration for how it is that the functionality within the math is actually connected. Maybe the command line interface will link it, and it should run correctly. As always, there's probably still work to be done.