v0.052f - Trading
I'm attempting to fix the actual training implementation and automatic triggers for the Linux pipeline in conjunction with CCXT triggering and other things that are needed, but I have to go step-by-step extremely slowly to ensure that I don't break functionality on Linux.

We're just re-importing correct functionality with our main core using backups.