v0.052k - Strategy Integration
We're attempting to essentially re-implement a whole bunch of backup files.

The point of these is to properly implement with entropy strategy, memory, etc. The things that would essentially be needed to continue to ensure functionality of the algorithmic trading bot.

Again, I shouldn't have to explain the math because that's pretty much the entire point of the project. If you're actually interested in the math, feel free to look at the code.