Update LICENSE.txt
This system is not intended for public use in any way, shape, or form. Failure to meet the requirements of the author will result in the removal of the entire repo. You my privately ruminate over the code inside your own head and imagine better code after reading or becoming attenuated with the code itself, this is provided for no other reason other than educational, while it still remains "up."